add unit tests for dependencies
