+++ pkg/model/schema_test.go
@@ -182,14 +182,13 @@ func Test_getStructKeys(t *testing.T) {
 			name:  "okteto manifest",
 			input: Manifest{},
 			expected: map[string][]string{
-				"dependencies.Dependency":    {"repository", "manifest", "branch", "wait", "timeout", "namespace"},
+				"dependencies.Dependency":    {"repository", "manifest", "branch", "namespace", "timeout", "wait"},
 				"env.Var":                    {"name", "value"},
 				"forward.Forward":            {"labels", "name", "localPort", "remotePort"},
 				"forward.GlobalForward":      {"labels", "name", "localPort", "remotePort"},
 				"model.BuildInfo":            {"secrets", "name", "context", "dockerfile", "target", "image", "cache_from", "export_cache", "depends_on"},
 				"model.Capabilities":         {"add", "drop"},
 				"model.ComposeInfo":          {"file", "services"},
-				"model.Dependency":           {"repository", "manifest", "branch", "namespace", "timeout", "wait"},
 				"model.DeployCommand":        {"name", "command"},
 				"model.DeployInfo":           {"endpoints", "image", "remote"},
 				"model.DestroyInfo":          {"image", "remote"},
